+++ packages/api/lib/writeRecords/utils.js
@@ -23,6 +23,7 @@ const {
 } = require('@cumulus/message/Providers');
 
 /**
+* @typedef {import('knex').Knex} Knex
 * @typedef {import('@cumulus/db').PostgresCollectionRecord} PostgresCollectionRecord
 **/
 
@@ -72,7 +73,7 @@ const getAsyncOperationCumulusId = async (
   }
 };
 
-const getParentExecutionCumulusId = async (
+const getParentExecution = async (
   parentExecutionArn,
   knex,
   executionPgModel = new ExecutionPgModel()
@@ -82,11 +83,12 @@ const getParentExecutionCumulusId = async (
       log.info('There is no parent execution ARN to lookup on the message, skipping');
       return undefined;
     }
-    return await executionPgModel.getRecordCumulusId(
+    return await executionPgModel.get(
       knex,
       {
         arn: parentExecutionArn,
-      }
+      },
+      ['cumulus_id', 'created_at']
     );
   } catch (error) {
     if (isFailedLookupError(error)) {
@@ -176,14 +178,14 @@ const getMessageProviderCumulusId = async (
 };
 
 /**
- * Looks up an Execution's cumulus_id by executionUrl.
+ * Looks up an Execution by executionUrl.
  *
  * @param {string} [executionUrl = ''] - Full url of stepfunction execution
  * @param {Knex} knex - knex Client
  * @param {Object} executionPgModel - instance of the exection database model
- * @returns {integer|undefined} - RDS internal cumulus_id
+ * @returns {Promise<import('@cumulus/db').PostgresExecutionRecord>} The returned record
  */
-const getExecutionCumulusId = async (
+const getExecution = async (
   executionUrl = '',
   knex,
   executionPgModel = new ExecutionPgModel()
@@ -193,9 +195,10 @@ const getExecutionCumulusId = async (
       log.info('There is no execution URL to lookup, skipping');
       return undefined;
     }
-    return await executionPgModel.getRecordCumulusId(
+    return await executionPgModel.get(
       knex,
-      { url: executionUrl }
+      { url: executionUrl },
+      ['cumulus_id', 'created_at']
     );
   } catch (error) {
     if (isFailedLookupError(error)) {
@@ -229,8 +232,8 @@ const isStatusActiveState = (status) => status === 'running' || status === 'queu
 module.exports = {
   isPostRDSDeploymentExecution,
   getAsyncOperationCumulusId,
-  getExecutionCumulusId,
-  getParentExecutionCumulusId,
+  getExecution,
+  getParentExecution,
   getProviderCumulusId,
   getCollectionCumulusId,
   getMessageProviderCumulusId,