+++ packages/api/tests/lib/writeRecords/test-write-granules.js
@@ -3,6 +3,7 @@
 const orderBy = require('lodash/orderBy');
 const test = require('ava');
 const cryptoRandomString = require('crypto-random-string');
+const { randomInt } = require('crypto');
 const sinon = require('sinon');
 const omit = require('lodash/omit');
 const range = require('lodash/range');
@@ -113,6 +114,7 @@ const updateGranule = async (t, updateGranulePayload, granuleWriteVia = 'api') =
   const {
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     granulePgModel,
     providerCumulusId,
@@ -139,6 +141,7 @@ const updateGranule = async (t, updateGranulePayload, granuleWriteVia = 'api') =
     await writeGranulesFromMessage({
       cumulusMessage: updatedCumulusMessage,
       executionCumulusId,
+      executionCreatedAt,
       providerCumulusId,
       knex,
       testOverrides: { stepFunctionUtils: t.context.stepFunctionUtils },
@@ -262,6 +265,7 @@ test.beforeEach(async (t) => {
     execution
   );
   t.context.executionCumulusId = pgExecution.cumulus_id;
+  t.context.executionCreatedAt = pgExecution.created_at;
   t.context.executionUrl = pgExecution.url;
 
   const [pgProvider] = await t.context.providerPgModel.create(
@@ -365,10 +369,11 @@ test.serial('getGranuleFromQueryResultOrLookup() returns cumulus ID from databas
 test.serial('writeFilesViaTransaction() throws error if any writes fail', async (t) => {
   const { knex } = t.context;
 
-  const fileRecords = [
-    fakeFileRecordFactory(),
-    fakeFileRecordFactory(),
-  ];
+  const fileRecords = range(2).map(() =>
+    fakeFileRecordFactory({
+      granule_cumulus_id: randomInt(10),
+      collection_cumulus_id: randomInt(10),
+    }));
 
   const fakeFilePgModel = {
     upsert: sinon.stub()
@@ -395,6 +400,7 @@ test.serial('_writeGranule will not allow a running status to replace a complete
   const {
     granule,
     executionCumulusId,
+    executionCreatedAt,
     collectionCumulusId,
     granuleId,
     granulePgModel,
@@ -413,6 +419,7 @@ test.serial('_writeGranule will not allow a running status to replace a complete
     apiGranuleRecord,
     postgresGranuleRecord,
     executionCumulusId,
+    executionCreatedAt,
     granulePgModel,
     knex,
     snsEventType: 'Update',
@@ -457,6 +464,7 @@ test.serial('_writeGranule will not allow a running status to replace a complete
     apiGranuleRecord: updatedapiGranuleRecord,
     postgresGranuleRecord: updatedPgGranuleRecord,
     executionCumulusId,
+    executionCreatedAt,
     granulePgModel,
     knex,
     snsEventType: 'Update',
@@ -480,6 +488,7 @@ test.serial('writeGranuleExecutionAssociationsFromMessage() saves granule-execut
       knex,
       collectionCumulusId,
       executionCumulusId,
+      executionCreatedAt,
       providerCumulusId,
       granuleId,
       stepFunctionUtils,
@@ -514,6 +523,7 @@ test.serial('writeGranuleExecutionAssociationsFromMessage() saves granule-execut
     await writeGranuleExecutionAssociationsFromMessage({
       cumulusMessage,
       executionCumulusId,
+      executionCreatedAt,
       knex,
     });
 
@@ -538,6 +548,7 @@ test.serial(
       knex,
       collectionCumulusId,
       executionCumulusId,
+      executionCreatedAt,
       granuleId,
     } = t.context;
 
@@ -554,6 +565,7 @@ test.serial(
     await writeGranuleExecutionAssociationsFromMessage({
       cumulusMessage,
       executionCumulusId,
+      executionCreatedAt,
       knex,
     });
 
@@ -572,15 +584,19 @@ test.serial(
       cumulusMessage,
       knex,
       executionCumulusId,
+      executionCreatedAt,
     } = t.context;
 
     const granuleCount = 5;
     const failedGranuleCount = 3;
 
-    const granuleCumulusIds = range(granuleCount).map(() =>
-      Math.floor(Math.random() * 100) + 1);
-    const getStub = sinon.stub(GranulePgModel.prototype, 'getRecordsCumulusIds')
-      .callsFake(() => granuleCumulusIds);
+    const pgGranules = range(granuleCount).map(() => ({
+      cumulus_id: Math.floor(Math.random() * 100) + 1,
+      collection_cumulus_id: 1,
+    }));
+
+    const getStub = sinon.stub(GranulePgModel.prototype, 'getRecords')
+      .callsFake(() => pgGranules);
 
     const errorMessage = 'fail';
     let count = 0;
@@ -604,6 +620,7 @@ test.serial(
       writeGranuleExecutionAssociationsFromMessage({
         cumulusMessage,
         executionCumulusId,
+        executionCreatedAt,
         knex,
       })
     );
@@ -625,6 +642,7 @@ test.serial('writeGranulesFromMessage() returns undefined if message has no gran
   const {
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     stepFunctionUtils,
   } = t.context;
@@ -632,6 +650,7 @@ test.serial('writeGranulesFromMessage() returns undefined if message has no gran
   const actual = await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -643,6 +662,7 @@ test.serial('writeGranulesFromMessage() returns undefined if message has empty g
   const {
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     stepFunctionUtils,
   } = t.context;
@@ -650,6 +670,7 @@ test.serial('writeGranulesFromMessage() returns undefined if message has empty g
   const actual = await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -663,6 +684,7 @@ test.serial('writeGranulesFromMessage() saves granule records to PostgreSQL/SNS'
     knex,
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -674,6 +696,7 @@ test.serial('writeGranulesFromMessage() saves granule records to PostgreSQL/SNS'
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -697,6 +720,7 @@ test.serial('writeGranulesFromMessage() propagates producerGranuleId with value 
     knex,
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     granulePgModel,
@@ -711,6 +735,7 @@ test.serial('writeGranulesFromMessage() propagates producerGranuleId with value 
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -734,6 +759,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     executionUrl,
     files,
     granulePgModel,
@@ -800,6 +826,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -819,6 +846,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -877,6 +905,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     executionUrl,
     files,
     granulePgModel,
@@ -899,6 +928,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -919,6 +949,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -942,6 +973,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     executionUrl,
     files,
     granulePgModel,
@@ -987,6 +1019,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1017,6 +1050,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1067,6 +1101,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -1129,6 +1164,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1166,6 +1202,7 @@ test.serial('writeGranulesFromMessage() on re-write saves granule records to Pos
     providerCumulusId,
     knex,
     executionCumulusId: pgExecution.cumulus_id,
+    executionCreatedAt: pgExecution.created_at,
     testOverrides: { stepFunctionUtils },
   });
 
@@ -1207,6 +1244,7 @@ test.serial('writeGranulesFromMessage() sets a default value of false for `publi
     cumulusMessage,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -1218,6 +1256,7 @@ test.serial('writeGranulesFromMessage() sets a default value of false for `publi
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1244,6 +1283,7 @@ test.serial('writeGranulesFromMessage() uses a default value for granule.created
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -1257,6 +1297,7 @@ test.serial('writeGranulesFromMessage() uses a default value for granule.created
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1282,6 +1323,7 @@ test.serial('writeGranulesFromMessage() allows overwrite of createdAt and uses g
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -1291,6 +1333,7 @@ test.serial('writeGranulesFromMessage() allows overwrite of createdAt and uses g
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1319,6 +1362,7 @@ test.serial('writeGranulesFromMessage() given a payload with undefined files, ke
     granulePgModel,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -1344,6 +1388,7 @@ test.serial('writeGranulesFromMessage() given a payload with undefined files, ke
   await writeGranulesFromMessage({
     cumulusMessage: completedCumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1403,6 +1448,7 @@ test.serial('writeGranulesFromMessage() given a partial granule overwrites only 
     granulePgModel,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -1429,6 +1475,7 @@ test.serial('writeGranulesFromMessage() given a partial granule overwrites only 
   await writeGranulesFromMessage({
     cumulusMessage: completedCumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1481,6 +1528,7 @@ test.serial('writeGranulesFromMessage() given an empty array as a files key will
   const {
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granule,
     granuleId,
@@ -1510,6 +1558,7 @@ test.serial('writeGranulesFromMessage() given an empty array as a files key will
   await writeGranulesFromMessage({
     cumulusMessage: completedCumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1572,6 +1621,7 @@ test.serial('writeGranulesFromMessage() given a null files key will throw an err
     granule,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -1597,6 +1647,7 @@ test.serial('writeGranulesFromMessage() given a null files key will throw an err
   await writeGranulesFromMessage({
     cumulusMessage: completedCumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1641,6 +1692,7 @@ test.serial('writeGranulesFromMessage() removes preexisting granule file from Po
   const {
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     filePgModel,
     granule,
     granulePgModel,
@@ -1675,6 +1727,7 @@ test.serial('writeGranulesFromMessage() removes preexisting granule file from Po
   await Promise.all(files.map(async (file) => {
     const pgFile = await translateApiFiletoPostgresFile(file);
     pgFile.granule_cumulus_id = existingPgGranuleRecord.cumulus_id;
+    pgFile.collection_cumulus_id = existingPgGranuleRecord.collection_cumulus_id;
     return filePgModel.create(knex, pgFile);
   }));
   const existingPgFiles = await filePgModel.search(knex, {});
@@ -1690,6 +1743,7 @@ test.serial('writeGranulesFromMessage() removes preexisting granule file from Po
 
   const [fakeFile] = await filePgModel.create(knex, {
     granule_cumulus_id: returnedGranule[0].cumulus_id,
+    collection_cumulus_id: returnedGranule[0].collection_cumulus_id,
     bucket: 'fake_bucket',
     key: 'fake_key',
   }, '*');
@@ -1700,6 +1754,7 @@ test.serial('writeGranulesFromMessage() removes preexisting granule file from Po
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     knex,
     testOverrides: { stepFunctionUtils },
   });
@@ -1725,6 +1780,7 @@ test.serial('writeGranulesFromMessage() saves file records to when workflow stat
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     filePgModel,
     files,
     granuleId,
@@ -1739,6 +1795,7 @@ test.serial('writeGranulesFromMessage() saves file records to when workflow stat
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1774,6 +1831,7 @@ test.serial('writeGranulesFromMessage() handles successful and failing writes in
     knex,
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     stepFunctionUtils,
   } = t.context;
@@ -1789,6 +1847,7 @@ test.serial('writeGranulesFromMessage() handles successful and failing writes in
   await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     knex,
     testOverrides: { stepFunctionUtils },
   }));
@@ -1806,6 +1865,7 @@ test.serial('writeGranulesFromMessage() throws error if any granule writes fail'
     cumulusMessage,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     stepFunctionUtils,
   } = t.context;
@@ -1819,6 +1879,7 @@ test.serial('writeGranulesFromMessage() throws error if any granule writes fail'
   await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1830,6 +1891,7 @@ test.serial('writeGranulesFromMessage() does not write to PostgreSQL/SNS if Post
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -1847,6 +1909,7 @@ test.serial('writeGranulesFromMessage() does not write to PostgreSQL/SNS if Post
   const [error] = await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granulePgModel: testGranulePgModel,
     knex,
     providerCumulusId,
@@ -1875,6 +1938,7 @@ test.serial('writeGranulesFromMessage() writes a granule and marks as failed if 
     knex,
     collectionCumulusId,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -1888,6 +1952,7 @@ test.serial('writeGranulesFromMessage() writes a granule and marks as failed if 
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1908,6 +1973,7 @@ test.serial('writeGranuleFromMessage() writes a new granule with files set to "[
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     granulePgModel,
     knex,
@@ -1920,6 +1986,7 @@ test.serial('writeGranuleFromMessage() writes a new granule with files set to "[
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1943,6 +2010,7 @@ test.serial('_writeGranules attempts to mark granule as failed if a SchemaValida
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -1955,6 +2023,7 @@ test.serial('_writeGranules attempts to mark granule as failed if a SchemaValida
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     knex,
     testOverrides: { stepFunctionUtils },
   });
@@ -1975,6 +2044,7 @@ test.serial('_writeGranules attempts to mark granule as failed if a SchemaValida
   const [error] = await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -1992,6 +2062,7 @@ test.serial('writeGranulesFromMessage() does not write any files if there is an 
   const {
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     filePgModel,
     granulePgModel,
     knex,
@@ -2017,6 +2088,7 @@ test.serial('writeGranulesFromMessage() does not write any files if there is an 
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2038,6 +2110,7 @@ test.serial('writeGranulesFromMessage() stores error on granule if any file fail
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -2062,6 +2135,7 @@ test.serial('writeGranulesFromMessage() stores error on granule if any file fail
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2081,6 +2155,7 @@ test.serial('writeGranulesFromMessage() stores an aggregate workflow error and f
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -2095,6 +2170,7 @@ test.serial('writeGranulesFromMessage() stores an aggregate workflow error and f
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2115,6 +2191,7 @@ test.serial('writeGranulesFromMessage() honors granule.createdAt time if provide
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -2128,6 +2205,7 @@ test.serial('writeGranulesFromMessage() honors granule.createdAt time if provide
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2145,6 +2223,7 @@ test.serial('writeGranulesFromMessage() throws if workflow_start_time is not pro
     cumulusMessage,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     stepFunctionUtils,
   } = t.context;
@@ -2154,6 +2233,7 @@ test.serial('writeGranulesFromMessage() throws if workflow_start_time is not pro
   await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2165,6 +2245,7 @@ test.serial('writeGranulesFromMessage() falls back to workflow_start_time if gra
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     knex,
     providerCumulusId,
@@ -2180,6 +2261,7 @@ test.serial('writeGranulesFromMessage() falls back to workflow_start_time if gra
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2198,6 +2280,7 @@ test.serial('writeGranulesFromMessage() sets `published` to false if null value 
     cumulusMessage,
     knex,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     granuleId,
     stepFunctionUtils,
@@ -2209,6 +2292,7 @@ test.serial('writeGranulesFromMessage() sets `published` to false if null value 
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2235,6 +2319,7 @@ test.serial('writeGranulesFromMessage() does not write a granule to Postgres if 
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     granuleId,
     granulePgModel,
     knex,
@@ -2260,6 +2345,7 @@ test.serial('writeGranulesFromMessage() does not write a granule to Postgres if 
   const [error] = await t.throwsAsync(writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     granulePgModel,
@@ -2288,6 +2374,7 @@ test.serial('writeGranulesFromMessage() does not persist file records to Postgre
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     filePgModel,
     granuleId,
     granulePgModel,
@@ -2301,6 +2388,7 @@ test.serial('writeGranulesFromMessage() does not persist file records to Postgre
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2325,6 +2413,7 @@ test.serial('writeGranulesFromMessage() does not persist file records to Postgre
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     filePgModel,
     granuleId,
     granulePgModel,
@@ -2338,6 +2427,7 @@ test.serial('writeGranulesFromMessage() does not persist file records to Postgre
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2366,6 +2456,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     executionUrl,
     files,
     granulePgModel,
@@ -2388,6 +2479,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2414,6 +2506,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2479,6 +2572,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     executionUrl,
     files,
     granulePgModel,
@@ -2501,6 +2595,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2532,6 +2627,7 @@ test.serial('writeGranulesFromMessage() on re-write with the same granule values
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2591,6 +2687,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -2611,6 +2708,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2654,6 +2752,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2704,6 +2803,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -2724,6 +2824,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2767,6 +2868,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2816,6 +2918,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -2836,6 +2939,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2888,6 +2992,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId: pgExecution.cumulus_id,
+    executionCreatedAt: pgExecution.created_at,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -2942,6 +3047,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -2962,6 +3068,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3014,6 +3121,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId: pgExecution.cumulus_id,
+    executionCreatedAt: pgExecution.created_at,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3062,6 +3170,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -3082,6 +3191,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3133,6 +3243,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "ru
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId: pgExecution.cumulus_id,
+    executionCreatedAt: pgExecution.created_at,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3181,6 +3292,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -3201,6 +3313,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3252,6 +3365,7 @@ test.serial('writeGranulesFromMessage() on update changing granule status to "qu
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId: pgExecution.cumulus_id,
+    executionCreatedAt: pgExecution.created_at,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3300,6 +3414,7 @@ test.serial('writeGranulesFromMessage() on update with "completed" status and st
     collectionCumulusId,
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     files,
     granulePgModel,
     knex,
@@ -3320,6 +3435,7 @@ test.serial('writeGranulesFromMessage() on update with "completed" status and st
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3366,6 +3482,7 @@ test.serial('writeGranulesFromMessage() on update with "completed" status and st
   await writeGranulesFromMessage({
     cumulusMessage,
     executionCumulusId,
+    executionCreatedAt,
     providerCumulusId,
     knex,
     testOverrides: { stepFunctionUtils },
@@ -3425,6 +3542,7 @@ test.serial('writeGranuleFromApi() removes preexisting granule file from postgre
 
   const [fakeFile] = await filePgModel.create(knex, {
     granule_cumulus_id: returnedGranule[0].cumulus_id,
+    collection_cumulus_id: returnedGranule[0].collection_cumulus_id,
     bucket: 'fake_bucket',
     key: 'fake_key',
   }, '*');
@@ -4357,7 +4475,7 @@ test.serial('updateGranuleStatusToQueued() updates granule status in PostgreSQL 
   const { Messages } = await sqs().receiveMessage({
     QueueUrl,
     MaxNumberOfMessages: 2,
-    WaitTimeSeconds: 10,
+    WaitTimeSeconds: 20,
   });
   const snsMessageBody = JSON.parse(Messages[1].Body);
   const publishedMessage = JSON.parse(snsMessageBody.Message);
@@ -4462,6 +4580,7 @@ test.serial('_writeGranule() successfully publishes an SNS message', async (t) =
   const {
     granule,
     executionCumulusId,
+    executionCreatedAt,
     knex,
     granulePgModel,
     granuleId,
@@ -4481,6 +4600,7 @@ test.serial('_writeGranule() successfully publishes an SNS message', async (t) =
     apiGranuleRecord,
     postgresGranuleRecord,
     executionCumulusId,
+    executionCreatedAt,
     granulePgModel,
     knex,
     snsEventType: 'Update',